Set both TCcountArgs funcs to size_t and fixed warnings NOTE \/ \/ \/
Leaving 2 types warnings in: 1 TBmakeNum(uint), 2 Linksign/NUMPARAMS

These will be used later when the AST is changed to store them as size_t

Linksign and NUMPARAMS are linked with TCcountArgs and Rets and
these will be left until it is known for certain that they never become
negative as it would be unsafe to assume this and set them to size_t

+++ src/libsac2c/tree/tree_basic.c
@@ -267,10 +267,10 @@ TBfreeCudaAccessInfo (cuda_access_info_t *access_info)
 /*--------------------------------------------------------------------------*/
 
 argtab_t *
-TBmakeArgtab (int size)
+TBmakeArgtab (size_t size)
 {
     argtab_t *argtab;
-    int i;
+    size_t i;
 
     DBUG_ENTER ();
